Differentiate between input and output attribute
The `claim_name` is the name in the _output_ QDL stem that preserves
what name was originally searched for.  It should always be different
from the LDAP attribute you are interested in for the result.  So,
`claim_name` should never show up in the list of search attributes.

OA4MP has special handling logic that deletes the `claim_name` attribute
if the attribute was found in LDAP and it matches the original name
being searched for, resulting in surprising behavior without this change.